Create Task56.java
The robot can move in four directions ("11" - north, "12" - west, "13" - south, "14" - east) and receive three numeric commands: 0 - keep moving, 1 - turn left, -1 - turn right. Given a number (11, 12, 13 or 14) - the initial direction of the robot and an integer N (0, 1 or -1) - the command sent to it. Output the direction of the robot after the command is received (i.e. north, west, south or east).