Create Task73.java
Given two rectangles whose sides are parallel or perpendicular to the coordinate axes. The coordinates of the bottom left corner of each rectangle and the lengths of their sides are known. Call one of the rectangles the first, the other the second.
a) Determine whether all the points of the first rectangle belong to the second rectangle.
b) Determine whether all points of one rectangle belong to the other rectangle.
c) Determine whether these rectangles intersect.