Create Task84.java
The user enters the dollar exchange rate in roubles. Show a table with the prices of $1, $2, ..., $100 in roubles, and in the third column add the number of candies you can buy with these amounts, if the price of 1 kg of candy is 20 roubles. Example: 1$ - 70p - 3.5kg and so on (100 rows in total).